pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pocketguide"
version = "0.1.0"
description = "Offline-capable travel assistant LLM with evaluation-first design"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "PocketGuide Team"}
]
dependencies = [
    "pyyaml==6.0.1",
    "transformers==4.46.3",
    "torch==2.5.1",
    "peft==0.14.0",
    "accelerate==0.33.0",
    "jsonschema==4.21.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "ruff==0.1.9",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
